Car GameObject (Lesson 7)
Unity crashed when I was making the course in the previous lesson so I just used the final scene for the track and environment.  I removed all other elements in the scene to start in the right place.
This lesson was about adding and setting up the car and follow camera